buypage: rename MyOrder to MyOrderTab
